Obtain the list of public logs from Google
The RFC for certificate transparency does not specify a way to obtain
a log's public key from the API. Instead, you are supposed to obtain
it through other means. Fortunately, Google keeps a list of logs around
which one can search inside for a matching URL.

That way, users don't have to manually specify the key.
Maybe in the future one could add such manual ways, but not right now.